resourceUseStats moved to a diffrente file. Use() renamed to AddUseCount(). Several minor comments and fixes
